Extract version utils
